requirements/dev.txt
ipdb==0.7
ipython==0.13
django-debug-toolbar==0.9.4
